docs: jest testing shim refactor for re-usability
Our team is using this as a separate file in our jest project. We noticed it was being globally injected into our app. Best practice would be to load this as a separate file that can be run on demand. Otherwise you have to copy/paste this snippet into the global setup or add it to multiple files.

* Encapsulated the global shim to prevent accidentally hijacking a React app's DOM
* Added an init to make sure the shim is only applied once when the first `mockReactFlow()` call is made