Add searchable description of deleteKeyCode
We get a question at least once per month in the Discord server asking how to disable the button that deletes nodes. If a user currently searches for "backspace," this page doesn't appear as a result. Adding this short description under the Keys headline should make this section searchable and less people should have to troubleshoot this.